+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/admin/staff/CourtWorkerController.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ethos.replacement.docmosis.controllers.admin.staff.courtworker;
+package uk.gov.hmcts.ethos.replacement.docmosis.controllers.admin.staff;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -16,7 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.ethos.replacement.docmosis.domain.admin.CCDCallbackResponse;
 import uk.gov.hmcts.ethos.replacement.docmosis.domain.admin.CCDRequest;
 import uk.gov.hmcts.ethos.replacement.docmosis.service.VerifyTokenService;
-import uk.gov.hmcts.ethos.replacement.docmosis.service.admin.staff.courtworker.CourtWorkerService;
+import uk.gov.hmcts.ethos.replacement.docmosis.service.admin.staff.CourtWorkerService;
 
 import java.util.List;
